==> src/1 get_articles_cross_ref.ipynb <==
# Get List of articles usnig crossref api

### import the required libraries

import requests
import pandas as pd
from time import time

### Set API address and parameters

api_url = "https://api.crossref.org/works"
query_params = {
    "query": "LiDAR SLAM",
    "rows": 1000,  # Adjust based on how many results you want per page
    # "offset": 0, # Deprecated use cursor instead
    "cursor": "*"  # Starting cursor
}

### Make test request

test_params = query_params.copy()
test_params["rows"] = 10

response = requests.get(api_url, params=test_params)
if response.status_code != 200: raise SystemExit("Bad response")

data = response.json()
Number_of_articles = data['message']['total-results']
print(f"Query: {data['message']['query']}")
print(f"The number of work matched criteria = {Number_of_articles}\n")

### Function to get information about the article

import unicodedata

def preprocess_text(text):
    text = text.replace("\n", " ") # remove new lines
    text = text.replace('/',' / ') # ensure space between words
    text = text.replace(",", ", ").replace(",  ", ", ") # replace commas without space
    return unicodedata.normalize('NFKD', text) 

def get_article_info(item):
    Authors = item.get('author', [])
    article_info = {
        "title": preprocess_text(item.get('title', ['No Title'])[0]),
        "DOI": item.get('DOI'),
        "publisher": item.get('publisher'),
        "publisher_location": item.get('publisher-location'),
        "year": item.get('created', {}).get('date-parts', [[None]])[0][0],
        "type": item.get('type'),
        "reference-count": item.get('reference-count'),
        "is-referenced-by-count": item.get('is-referenced-by-count'),
        "authors": ', '.join([author.get('given', '') + " " + author.get('family', '') for author in Authors])
    }
    
    try: # use try because some fields may be empty
        first_author_affiliation = Authors[0].get('affiliation', [])
        affiliation_location = first_author_affiliation[0].get('name')
        article_info["publisher_affiliation"] = preprocess_text(affiliation_location)
    except Exception:
        article_info["publisher_affiliation"] = None
    
    return article_info

### Getting the list of articles and saving them into an array 

cur_query_params = query_params.copy()
articles = []  # List to store article information

start = time()
while True: 
    # make request
    response = requests.get(api_url, params=cur_query_params) 
    if response.status_code != 200: raise SystemExit(f"Bad response ({response.status_code}): {response.json()["message"]}")

    # read list of works    
    data = response.json()
    items = data['message']['items']

    # stop if no work remains
    if not items: break

    # get article info
    for item in items:
        articles.append(get_article_info(item))

    cur_query_params['cursor'] = data['message'].get('next-cursor')
    print(f"articles found: {len(articles)}/{Number_of_articles} ({(time() - start):.2f} sec)")

### Create pandas dataframe 

df = pd.DataFrame(articles)
df.iloc[:].head()

df.info()

### Save data

# Save the DataFrame to a CSV file with proper handling of newline characters
df.to_csv('slam_articles.csv', index=False, encoding='utf-8', quoting=1)
