==> src/2 get_country_by_address.ipynb <==
# Use the address and university to find the country

### import the required libraries

from geopy.geocoders import Nominatim
from time import time, sleep
import pandas as pd

### Declaring a list of countries for quick matching

# List of base countries
base_countries = {
    'Republic of Korea': 'South Korea',
    'Taiwan': 'China', 'Hong Kong': 'China', 'UAE': 'United Arab Emirates',
    'New York': 'United States', 'USA': 'United States', 'UK': 'United Kingdom', 
    
    'United States': 'United States', 'Germany': 'Germany', 'Italy': 'Italy', 
    'United Kingdom': 'United Kingdom', 'France': 'France', 'Canada': 'Canada', 
    'Brazil': 'Brazil', 'Australia': 'Australia', 'Spain': 'Spain', 'Russia': 'Russia', 
    'India': 'India', 'China': 'China', 'Japan': 'Japan', 'South Korea': 'South Korea', 
    'Mexico': 'Mexico', 'Netherlands': 'Netherlands', 'Switzerland': 'Switzerland', 
    'Sweden': 'Sweden', 'Belgium': 'Belgium', 'Austria': 'Austria', 'Norway': 'Norway', 
    'Denmark': 'Denmark', 'Finland': 'Finland', 'Greece': 'Greece', 'Portugal': 'Portugal', 
    'New Zealand': 'New Zealand', 'Ireland': 'Ireland', 'Poland': 'Poland', 
    'Czech Republic': 'Czech Republic', 'Turkey': 'Turkey', 'South Africa': 'South Africa', 
    'Argentina': 'Argentina', 'Hungary': 'Hungary', 'Chile': 'Chile', 'Singapore': 'Singapore', 
    'Israel': 'Israel', 'Thailand': 'Thailand', 'Malaysia': 'Malaysia', 'Indonesia': 'Indonesia', 
    'Ukraine': 'Ukraine', 'Vietnam': 'Vietnam', 'Philippines': 'Philippines', 'Egypt': 'Egypt', 
    'Saudi Arabia': 'Saudi Arabia', 'United Arab Emirates': 'United Arab Emirates', 
    'Pakistan': 'Pakistan', 'Iran': 'Iran', 'Iraq': 'Iraq', 'Kuwait': 'Kuwait', 'Qatar': 'Qatar', 
    'Nigeria': 'Nigeria', 'Kenya': 'Kenya', 'Morocco': 'Morocco', 'Algeria': 'Algeria', 
    'Tunisia': 'Tunisia', 'Ghana': 'Ghana', 'Ethiopia': 'Ethiopia', 'Cameroon': 'Cameroon', 
    'Uganda': 'Uganda', 'Senegal': 'Senegal', 'Zimbabwe': 'Zimbabwe', 'Zambia': 'Zambia', 
    'Tanzania': 'Tanzania', 'Botswana': 'Botswana', 'Namibia': 'Namibia', 'Mauritius': 'Mauritius', 
    'Madagascar': 'Madagascar', 'Angola': 'Angola', 'Mozambique': 'Mozambique', 'Rwanda': 'Rwanda', 
    'Ivory Coast': 'Ivory Coast', 'Burkina Faso': 'Burkina Faso', 'Uzbekistan': 'Uzbekistan', 
    'Kazakhstan': 'Kazakhstan', 'Turkmenistan': 'Turkmenistan', 'Azerbaijan': 'Azerbaijan', 
    'Georgia': 'Georgia', 'Armenia': 'Armenia', 'Belarus': 'Belarus', 'Moldova': 'Moldova', 
    'Kyrgyzstan': 'Kyrgyzstan', 'Tajikistan': 'Tajikistan', 'Latvia': 'Latvia', 
    'Lithuania': 'Lithuania', 'Estonia': 'Estonia', 'Slovakia': 'Slovakia', 'Slovenia': 'Slovenia', 
    'Croatia': 'Croatia', 'Bulgaria': 'Bulgaria', 'Romania': 'Romania', 'Serbia': 'Serbia', 
    'Bosnia and Herzegovina': 'Bosnia and Herzegovina', 'Montenegro': 'Montenegro', 
    'North Macedonia': 'North Macedonia', 'Albania': 'Albania', 'Kosovo': 'Kosovo', 
    'Cyprus': 'Cyprus', 'Malta': 'Malta', 'Iceland': 'Iceland', 'Luxembourg': 'Luxembourg', 
    'Liechtenstein': 'Liechtenstein', 'Monaco': 'Monaco', 'Andorra': 'Andorra', 'San Marino': 'San Marino', 
    'Vatican': 'Vatican'
}

### Define function to find country

from geopy.exc import GeocoderUnavailable

def get_country_from_list(address):
    for key_country, country in base_countries.items():
        if key_country.lower() in address.lower():
            return country #+ "*"
    return None

def get_country(address):
    country = get_country_from_list(address)
    if country: return country

    # If not, use geocoder to get the cosuntry
    geolocator = Nominatim(user_agent="address_to_country_converter")
    for _ in range(3): # Number of retries
        try:
            location = geolocator.geocode(address, language="en")
            return location.address.split(",")[-1].strip() if location else None
        except GeocoderUnavailable:
            # If geocoder is unavailable, wait and then retry
            print("Sleep")
            sleep(1)
    print("Service unavailable")
    return None

## Read dataset

import unicodedata

df = pd.read_csv('slam_articles.csv')
publishers = set(df["publisher"])
publisher_locations = set(df["publisher_location"])
universities = set(df["publisher_affiliation"])

# fix string text:
publishers = {str(pub) for pub in publishers if pub}
publisher_locations = {str(loc) for loc in publisher_locations if loc}
universities = {str(uni) for uni in universities if uni}

# remove bad values
for cur_set in [publishers, publisher_locations, universities]:
    cur_set.discard('nan')
    cur_set.discard(',')
    cur_set.discard('Unknown')
    cur_set.discard('unknown')

print(f"publishers count: {len(publishers)}")
print(f"publisher locations count: {len(publisher_locations)}")
print(f"universities count: {len(universities)}")

### Find countries for publishers (Use only base list)

publishers_dict = {}

N = len(publishers)
start = time()
for i, locat in enumerate(publishers):
    if i % 20 == 0:
        print(f"\r{i}/{N} ({(100*i/N):.1f}%): {(time() - start):.3f} sec ",end='')
    publishers_dict[locat] = get_country_from_list(locat)
print(f"\r{N}/{N} (100.0%): {(time() - start):.3f} sec ")

for pub, loc in list(publishers_dict.items())[10:15]:
    print(f"{pub}:\n\t{loc if loc else '-'}")

### Find countries for publisher_locations

publisher_locations_dict = {}

N = len(publisher_locations)
start = time()
for i, locat in enumerate(publisher_locations):
    if i % 20 == 0:
        print(f"\r{i}/{N} ({(100*i/N):.1f}%): {(time() - start):.3f} sec ",end='')
    publisher_locations_dict[locat] = get_country(locat)
print(f"\r{N}/{N} (100.0%): {(time() - start):.3f} sec ")

for pub, loc in list(publisher_locations_dict.items())[:5]:
    print(f"{pub}:\n\t{loc if loc else '-'}")

### Find countries for universities

universities_dict = {}

N = len(universities)
start = time()
for i, locat in enumerate(universities):
    if i % 20 == 0:
        print(f"\r{i}/{N} ({(100*i/N):.1f}%): {(time() - start):.3f} sec ",end='')
    universities_dict[locat] = get_country(locat)
print(f"\r{N}/{N} (100.0%): {(time() - start):.3f} sec ")

for pub, loc in list(universities_dict.items())[45:50]:
    print(f"{pub}:\n\t{loc if loc else '-'}")

### Check not None countries

not_none = 0
for location in universities_dict.values():
    if location: not_none+=1

print(not_none,'/',len(universities_dict),f"({(100*not_none/len(universities_dict)):.1f}%)")

print(set(publishers_dict.values()))
print(set(publisher_locations_dict.values()))
print(set(universities_dict.values()))

### Save to file

import json

with open('publishers_dict.json', "w") as file:
    json.dump(publishers_dict, file)

with open('publisher_locations_dict.json', "w") as file:
    json.dump(publisher_locations_dict, file)

with open('universities_dict.json', "w") as file:
    json.dump(universities_dict, file)
