==> src/3 add_country_to_dataset.ipynb <==
# Add new country column to articles database

import pandas as pd
from time import time

### Read original database

df = pd.read_csv('slam_articles.csv')

### Read location dictionaries

import json

with open('publishers_dict.json') as file:
    publishers_dict = json.loads(file.read())

with open('publisher_locations_dict.json') as file:
    publisher_locations_dict = json.loads(file.read())

with open('universities_dict.json') as file:
    universities_dict = json.loads(file.read())

print('publishers_dict:',len(publishers_dict))
print('publisher_locations_dict:',len(publisher_locations_dict))
print('universities_dict:',len(universities_dict))

### Create new columns

df['pub_c'] = None
df['pub_loc_c'] = None
df['uni_c'] = None
df.head()

### Fill new column

# add publisher
for key, value in publishers_dict.items():
    df.loc[df['publisher'] == key, "pub_c"] = value

# add publisher location
for key, value in publisher_locations_dict.items():
    df.loc[df['publisher_location'] == key, "pub_loc_c"] = value

# add university
for key, value in universities_dict.items():
    df.loc[df['publisher_affiliation'] == key, "uni_c"] = value

df.head()

### Combine country columns with prioritization

# Create the 'country' column with prioritization
df['country'] = df.apply(lambda row: row['uni_c'] if pd.notna(row['uni_c']) 
                               else (row['pub_loc_c'] if pd.notna(row['pub_loc_c']) 
                               else row['pub_c']), axis=1)
df = df.drop(columns=['pub_c', 'pub_loc_c', 'uni_c'])

df.head()

### Save new database

# Save the DataFrame to a CSV file with proper handling of newline characters
df.to_csv('slam_articles_with_country.csv', index=False, encoding='utf-8', quoting=1)

### Check how many articles have a country

df_filtered = df.dropna(subset=['country'])
N = len(df)
c_N = len(df_filtered)

print(f"{c_N}/{N} ({(100*c_N/N):.1f}%) articles with country")
